refactor(router-lite): moved the VIT generation from router to RouteContext
